+++ src/pages/EventDetails/components/EventRenderer.tsx
@@ -3,7 +3,7 @@ import classnames from 'classnames';
 import { Event } from 'types';
 import { PermissionApp } from 'types/Enums';
 import URLS from 'URLS';
-import { parseISO, isPast, isFuture, sub } from 'date-fns';
+import { parseISO, isPast, isFuture, subHours } from 'date-fns';
 import { formatDate, getICSFromEvent } from 'utils';
 import { Link } from 'react-router-dom';
 
@@ -171,33 +171,27 @@ const EventRenderer = ({ data, preview = false }: EventRendererProps) => {
         )}
         {(isFuture(signOffDeadlineDate) || registration.is_on_wait) && isFuture(startDate) ? (
           <VerifyDialog
-            closeText='Avbryt'
-            confirmText='Ja, jeg er sikker'
             contentText={`Om du melder deg på igjen vil du havne på bunnen av en eventuell venteliste.`}
             fullWidth
             onConfirm={signOff}
-            titleText='Er du sikker?'
             variant='outlined'>
             Meld deg av
           </VerifyDialog>
-        ) : isFuture(sub(Date.parse(data.start_date), { hours: 2 })) ? (
+        ) : isFuture(subHours(parseISO(data.start_date), 2)) ? (
           <>
             <VerifyDialog
-              closeText='Avbryt'
-              confirmText='Ja, jeg er sikker'
-              contentText={`Om du melder deg av nå vil du ikke kunne melde deg på igjen. I tillegg vil du få en prikk.`}
+              contentText={`Om du melder deg av nå vil du ikke kunne melde deg på igjen. Du vil også få 1 prikk for å melde deg av etter avmeldingsfristen.`}
               fullWidth
               onConfirm={signOff}
-              titleText='Er du sikker?'
               variant='outlined'>
               Meld deg av
             </VerifyDialog>
-            <Alert severity='warning' variant='outlined'>
-              Avmeldingsfristen har passert. Hvis du melder deg av nå, vil du få 1 prikk.
+            <Alert severity='info' variant='outlined'>
+              Avmeldingsfristen har passert. Du kan allikevel melde deg av frem til 2 timer før arrangementsstart, men du vil da få 1 prikk.
             </Alert>
           </>
         ) : (
-          <Alert severity='warning' variant='outlined'>
+          <Alert severity='info' variant='outlined'>
             Du kan ikke lengre melde deg av arrangementet
           </Alert>
         )}